+++ src/ca/mcgill/sable/soot/coffi/exception_table_entry.java
@@ -117,21 +117,21 @@ class exception_table_entry {
    /** Code offset of start of code protected by this handler (inclusive).
     * @see exception_table_entry#start_inst
     */
-   public short start_pc;
+   public int start_pc;
    /** Code offset of end of code protected by this handler (exclusive).
     * @see exception_table_entry#end_inst
     */
-   public short end_pc;
+   public int end_pc;
    /** Code offset of actual exception handler for the specified code block.
     * @see exception_table_entry#handler_inst
     * @see exception_table_entry#b
     */
-   public short handler_pc;
+   public int handler_pc;
    /** Constant pool index of a CONSTANT_Class entry describing the exception
     * this handler handles; if 0, this handler catches all exceptions.
     * @see CONSTANT_Class
     */
-   public short catch_type;
+   public int catch_type;
    /** First Instruction object (after parsing) of code protected by this handler.
     * @see exception_table_entry#start_pc
     * @see Instruction
